requirements.txt
rumps>=0.4.0
pyobjc-framework-Quartz>=12.1
